+++ prepare/Learning05/Square.cs
@@ -1,12 +1,12 @@
 public class Square : Shape
 {
-    private int _side;
+    private double _side;
 
-    public Square(string color, int side) : base(color)
+    public Square(string color, double side) : base(color)
     {
         _side = side;
     }
-    public override int GetArea()
+    public override double GetArea()
     {
         return _side * _side;
     }
